New plugin: add-conversions
When migrating to TypeScript the vast majority of errors I experienced
were `TS2339: Property '{0}' does not exist on type '{1}'`.

These get fixed with the `ts-ignore` plugin, but `ts-ignore` should be
a very last resort to fixing errors. In this case there's a better
option available: add a conversion to `any` (or the appropriate `any`
alias).

Doing this is better because it is less whitespace-sensitive and
silences exactly one error. (`ts-ignore`/`ts-expect-error` silence
errors for an entire line, which can contain multiple errors.)

With an `any` alias such as `$TSFixMe`, the conversions can still
stick out and be addressed later.